% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar.R
\name{abind_simple}
\alias{abind_simple}
\title{Bind arrays along the third dimension (simple abind replacement)}
\usage{
abind_simple(...)
}
\description{
Bind arrays along the third dimension (simple abind replacement)
}
\keyword{internal}
