% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor.R
\name{.compute_sddr_logspline}
\alias{.compute_sddr_logspline}
\title{Compute the SDDR using logspline (internal)}
\usage{
.compute_sddr_logspline(draws, prior, null = 0)
}
\arguments{
\item{draws}{Numeric vector of posterior draws.}

\item{prior}{A \code{bvarnet_prior} object.}

\item{null}{Numeric scalar — null value.}
}
\value{
Named list with elements \code{BF01}, \code{post_density},
\code{prior_density}.
}
\description{
Compute the SDDR using logspline (internal)
}
\keyword{internal}
