% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.extract_u_draws}
\alias{.extract_u_draws}
\title{Extract all posterior draws of subject-level random effects \code{u}}
\usage{
.extract_u_draws(object)
}
\arguments{
\item{object}{A \code{bvarnet} object.}
}
\value{
A 4D array with dimensions
\code{[S, p, J, n_re]} where \code{S = n_iter * n_chains}.
}
\description{
Returns a 4D array with dimensions
\code{[draw, node, subject, re]} matching the Stan declaration
\code{array[p] matrix[J, n_re] u}.
}
\details{
CmdStan flattens \code{array[p] matrix[J, n_re] u} as
\code{u[node, subject, re]} in column-major order within each array
element, i.e. \code{u[1,1,1], u[1,1,2], ..., u[1,J,n_re], u[2,1,1], ...}.
}
\keyword{internal}
