% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor.R
\name{.is_half_prior}
\alias{.is_half_prior}
\title{Determine whether a parameter uses a half-prior}
\usage{
.is_half_prior(param_name)
}
\description{
Determine whether a parameter uses a half-prior
}
\keyword{internal}
