% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor.R
\name{.param_label}
\alias{.param_label}
\title{Produce human-readable labels for a Stan parameter name (internal)}
\usage{
.param_label(stan_name, nm)
}
\description{
Produce human-readable labels for a Stan parameter name (internal)
}
\keyword{internal}
