% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor.R
\name{.param_type}
\alias{.param_type}
\title{Determine the prior type for a Stan parameter name}
\usage{
.param_type(param_name)
}
\description{
Determine the prior type for a Stan parameter name
}
\keyword{internal}
