% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar.R
\name{.parse_family}
\alias{.parse_family}
\title{Parse and validate the family argument}
\usage{
.parse_family(family, y_cols)
}
\arguments{
\item{family}{Character scalar or vector.}

\item{y_cols}{Character vector of outcome names.}
}
\value{
Named character vector.
}
\description{
Returns a named character vector of length \code{p} with names from
\code{y_cols}.
}
\keyword{internal}
