% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.posterior_mean_u}
\alias{.posterior_mean_u}
\title{Posterior mean of subject-level random effects \code{u}}
\usage{
.posterior_mean_u(object)
}
\arguments{
\item{object}{A \code{bvarnet} object.}
}
\value{
A 3D array with dimensions \code{[p, J, n_re]}.
}
\description{
Returns a 3D array \code{[node, subject, re]} containing the
posterior mean of each \code{u[node, subject, re]} element.
}
\keyword{internal}
