% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar.R
\name{.rename_node_param}
\alias{.rename_node_param}
\title{Rename a per-node (p=1) Stan parameter name to full-model naming}
\usage{
.rename_node_param(name, j, family, n_fe_full)
}
\description{
Rename a per-node (p=1) Stan parameter name to full-model naming
}
\keyword{internal}
