% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_priors.R
\name{print.bvarnet_prior}
\alias{print.bvarnet_prior}
\title{Print a bvarnet_prior}
\usage{
\method{print}{bvarnet_prior}(x, ...)
}
\arguments{
\item{x}{A \code{bvarnet_prior} object.}

\item{...}{Passed to \code{format.bvarnet_prior()}.}
}
\value{
\code{x} invisibly.
}
\description{
Print a bvarnet_prior
}
