% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentlife.R
\docType{data}
\name{studentlife}
\alias{studentlife}
\title{StudentLife Data}
\format{
\subsection{\code{studentlife}}{

A data frame with 1912 rows and 72 columns:
}
}
\source{
https://openesmdata.org/datasets/0004_wang/

https://doi.org/10.1145/2632048.2632054
}
\usage{
studentlife
}
\description{
Assessing mental health, academic performance and behavioral trends of college students using smartphones.
Wang, R., Chen, F., Chen, Z., Li, T., Harari, G., Tignor, S., Zhou, X., Ben-Zeev, D., & Campbell, A. T. (2014).
StudentLife: assessing mental health, academic performance and behavioral trends of college students using smartphones.
Proceedings of the 2014 ACM International Joint Conference on Pervasive and Ubiquitous Computing, 3–14. https://doi.org/10.1145/2632048.2632054
}
\keyword{datasets}
