% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvarnet-package.R
\docType{package}
\name{bvarnet-package}
\alias{bvarnet}
\alias{bvarnet-package}
\title{bvarnet: Bayesian Estimation of Dynamic VAR Models using STAN}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bayesian estimation of multilevel Vector Autoregression (VAR) models using Stan. Supports Gaussian, Binary, and Ordinal (adjacent category) outcome variables with random effects and customizable priors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://flo1met.github.io/bvarnet/}
  \item \url{https://github.com/flo1met/bvarnet}
  \item Report bugs at \url{https://github.com/flo1met/bvarnet/issues}
}

}
\author{
\strong{Maintainer}: Florian Metwaly \email{f.j.metwaly@uva.nl} (\href{https://orcid.org/0009-0001-8357-7870}{ORCID}) [copyright holder]

}
\keyword{internal}
