% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvarnet-package.R
\docType{package}
\name{bvarnet-package}
\alias{bvarnet}
\alias{bvarnet-package}
\title{bvarnet: Bayesian Estimation of Dynamic VAR Models using STAN}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Florian Metwaly \email{f.j.metwaly@uva.nl}

}
\keyword{internal}
