% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_priors.R
\name{get_default_priors}
\alias{get_default_priors}
\title{Get the default prior specification for a given model family}
\usage{
get_default_priors(family)
}
\arguments{
\item{family}{One of \code{"bernoulli"}, \code{"ordinal"}, \code{"gaussian"}.}
}
\value{
A \code{bvarnet_priors} object.
}
\description{
A convenience wrapper around \code{set_priors()} for inspecting defaults.
}
