% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_priors.R
\name{print.bvarnet_priors}
\alias{print.bvarnet_priors}
\title{Print a bvarnet_priors specification}
\usage{
\method{print}{bvarnet_priors}(x, ...)
}
\arguments{
\item{x}{A \code{bvarnet_priors} object.}

\item{...}{Ignored.}
}
\value{
\code{x} invisibly.
}
\description{
Print a bvarnet_priors specification
}
